Dependency optimization, initial FavoritesViewModel unit tests
While writting the first tests I found some hard dependency the  FavoriteViewModel had on the SmsController
it made sense to abstract this class with interface and abstract classes and move the current code to Impl classes.

This way its easier to mock and get more clearer seperation of concerns, this of course is subject to changes.

The modified classes are:
  - SmsBroadcastReceiver to SmsBroadcastReceiverImpl
  - SmsController to SmsControllerImpl
As mentioned above, the original are an abstract class and interface respectivly.

For #29