Some more abstraction into different product flavors
Sms module now has a mock() variant for easier testing, this was cared to not be
possible to ship with mockRelease (check variant filtering in sms module build.gradle)

With this, I hope to get a better better abstraction hence, better enviroment for integration tests.
Notice how the main folder of the sms module keeps the interface which represents the contract to
outer layers (mobile-ui) in this case, slowly but surely this moves us closer to the onion/clean arch.

For #29
See also: #25, #38